+++ CHANGE_LOG.md
@@ -10,6 +10,9 @@ All notable changes to this project will be documented in this file.
 
 - Adding runtimeId as part of events send to identify calls from different tabs.
 
+### Fixed
+- Fixed adaptive card error color issue
+
 ## [1.8.2] - 2025-08-20
 
 ### Fixed