+++ CHANGE_LOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - Fixed adaptive card error color issue
+- Citations reference label styling issue
 
 ## [1.8.2] - 2025-08-20